pyproject.toml
[tool.poetry]
name = "neura"
version = "0.1.0"
description = "Neural Entity for Understanding, Reasoning, and Autonomy - An LLM-powered agent framework"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "neura"}]

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.0.15"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
chromadb = "^0.4.18"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.1"
beautifulsoup4 = "^4.12.2"
playwright = "^1.40.0"
rich = "^13.7.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = [] 